Major bug fixes.
Had a bug where new subjects with non-unique names or short names could
be created + change the existing subjects with their names. Now every
subject name + short name must be unique. Also changed the layout of the
ManageSubject Panel to GridbagLayout to sucessully use the scrollbars.
Next fix will be the "bug" that the font size might be to big for the
schedule panel.